+++ src/app/auth.guard.ts
@@ -5,7 +5,7 @@ import {
   RouterStateSnapshot,
   Router,
 } from '@angular/router';
-import { AuthService } from './services/auth.service';
+import { AuthService } from '@services/auth.service';
 
 @Injectable({
   providedIn: 'root',
@@ -21,6 +21,17 @@ export class AuthGuard implements CanActivate {
     state: RouterStateSnapshot,
   ): boolean {
     if (this.authService.isAuthenticatedBase()) {
+      if (route.routeConfig?.path === 'crew-area') {
+        if (this.authService.isAuthenticatedCrew()) {
+          return true;
+        } else {
+          alert('insufficient access rights for this route');
+          this.router.navigate(['/login'], {
+            queryParams: { redirect: state.url },
+          });
+          return false;
+        }
+      }
       return true;
     } else {
       this.router.navigate(['/login'], {